AArch64: Error checking of temporary file handling
Failures in the posix::FOpen call in the String GetCapturedString() method
of the CapturedStream class were ignored: the file variable was passed to
ReadEntireFile(file), causing a segmentation violation without indication
as to cause. In this commit GTEST_CHECK_ is now used to provide an indication
of the reason for failure.

Change-Id: Id419f4fa1f05b5c3ca09a19551747bb5f0ee635f
Signed-off-by: Marcus Oakland <marcus.oakland@arm.com>