Android.bp: Move from src&test to top level
Soong is enforcing that all relative file references do not escape the
directory that the Android.bp is located in. All of the code expects to
be built from the top level anyway (#include "src/..."), so just move
the build files up.

Also refactors the test modules to use test_per_src and match the
options used in the Android.mk.

Change-Id: I79651b0bd3569a2dfc07119db0415aa6ba9712e0